Enhancements: - Point system from API with rounding logic - If match started, block the user to bet on that match - displaying email id to usernames - Blocking pots less than 3 Bug Fixes

+++ SportPot/ManagedObjects/FixtureMO+CoreDataClass.swift
@@ -13,7 +13,7 @@ import CoreData
 public class FixtureMO: NSManagedObject, Decodable {
     var isDoubleDown: Bool = false
     var predictionType: PredictionType = .none
-    var selectedPoints: Double = 0.0
+    var selectedPoints: Int = 0
     
     private enum CodingKeys: String, CodingKey {
         case elapsed
